+++ src/openhems/modules/energy_strategy/__init__.py
@@ -5,6 +5,7 @@ Module for choosing a energy management strategy.
 
 from .energy_strategy import EnergyStrategy, LOOP_DELAY_VIRTUAL
 from .offpeak_strategy import OffPeakStrategy
+from .switchoff_strategy import SwitchoffStrategy
 # from .emhass_strategy import EmhassStrategy
 from .solarbased_strategy import SolarBasedStrategy
 from .hybridinverter_strategy import HybridInverterStrategy