+++ src/openhems/modules/energy_strategy/offpeak_strategy.py
@@ -5,13 +5,10 @@ This is in case we just base on "off-peak" range hours to control output.
 """
 
 from datetime import datetime
-import time
 from openhems.modules.network.network import OpenHEMSNetwork
 from openhems.modules.util import ConfigurationException
 from .energy_strategy import EnergyStrategy, LOOP_DELAY_VIRTUAL
 
-# Time to wait in seconds before considering to be in offpeak range
-TIME_MARGIN_IN_S = 1
 
 # pylint: disable=broad-exception-raised
 class OffPeakStrategy(EnergyStrategy):
@@ -21,20 +18,18 @@ class OffPeakStrategy(EnergyStrategy):
 	The strategy is to switch on electric devices only on "off-peak" hours
 	 with check to not exceed authorized max consumption
 	"""
-	offpeakHoursRanges = []
-	inOffpeakRange = False
-	rangeEnd = datetime.now()
-	network = None
 
-	def __init__(self, mylogger, network: OpenHEMSNetwork):
-		super().__init__(mylogger)
-		self.network = network
+	def __init__(self, mylogger, network: OpenHEMSNetwork, strategyId:str):
+		super().__init__(strategyId, network, mylogger, True)
+		self.inOffpeakRange = False
+		self.rangeEnd = datetime.now()
 		self.offpeakHoursRanges = self.network.getOffPeakHoursRanges()
-		self.logger.info("OffPeakStrategy(%s)", str(self.offpeakHoursRanges))
+		self.logger.info("OffPeakStrategy(%s) on %s", str(self.offpeakHoursRanges), str(self.getNodes()))
 		if not self.offpeakHoursRanges:
 			msg = "OffPeak-strategy is useless without offpeak hours. Check your configuration."
 			self.logger.critical(msg)
 			raise ConfigurationException(msg)
+		self._rangeChangeDone = False
 		self.checkRange()
 
 	def checkRange(self, nowDatetime: datetime=None) -> int:
@@ -43,19 +38,10 @@ class OffPeakStrategy(EnergyStrategy):
 		 and set end time of this range
 		"""
 		self.offpeakHoursRanges = self.network.getOffPeakHoursRanges()
+		inoffpeak = self.inOffpeakRange
 		self.inOffpeakRange, self.rangeEnd = self.offpeakHoursRanges.checkRange(nowDatetime)
-
-	def sleepUntillNextRange(self):
-		"""
-		Set application to sleep until off-peak (or inverse) range end
-		TIME_MARGIN_IN_S: margin to wait more to be sure to change range... 
-		useless, not scientist?
-		"""
-		time2wait = (self.rangeEnd - datetime.now()).total_seconds()
-		self.logger.info("OffPeakStrategy.sleepUntillNextRange() : "
-			"sleep(%d min, until %s)",\
-			round((time2wait+TIME_MARGIN_IN_S)/60), str(self.rangeEnd))
-		time.sleep(time2wait+TIME_MARGIN_IN_S)
+		if inoffpeak!=self.inOffpeakRange:
+			self._rangeChangeDone = False
 
 	def switchOnMax(self, cycleDuration):
 		"""
@@ -63,35 +49,43 @@ class OffPeakStrategy(EnergyStrategy):
 		 - If there is no margin to switch on, do nothing.
 		 - Only one (To be sure to not switch on to much devices)
 		"""
-		self.logger.info("OffPeakStrategy.switchOnMax()")
+		self.logger.info("%s.switchOnMax()", self.strategyId)
 		done = 0
 		marginPower = self.network.getMarginPowerOn()
 		doSwitchOn = True
 		if marginPower<0:
 			self.logger.info("Can't switch on devices: not enough power margin : %s", marginPower)
 			return True
-		for elem in self.network.getAll("out"):
-			if self.switchOn(elem, cycleDuration, doSwitchOn):
+		for elem in self.getNodes():
+			if self.switchOnSchedulable(elem, cycleDuration, doSwitchOn):
 				# Do just one at each loop to check Network constraint
 				doSwitchOn = False
 				done += 1
 		return done == 0
 
-	def updateNetwork(self, cycleDuration):
+	def updateNetwork(self, cycleDuration:int, allowSleep:bool, now=None) -> int:
 		"""
 		Decide what to do during the cycle:
 		 IF off-peak : switch on all
 		 ELSE : Switch off all AND Sleep until off-peak
 		"""
-		if datetime.now()>self.rangeEnd:
+		if now is None:
+			now = datetime.now()
+		if now>self.rangeEnd:
 			self.checkRange()
 		if self.inOffpeakRange:
 			# We are in off-peak range hours : switch on all
 			self.switchOnMax(cycleDuration)
 		else: # Sleep untill end.
-			if self.network.switchOffAll():
-				if cycleDuration>LOOP_DELAY_VIRTUAL:
-					self.sleepUntillNextRange()
-					self.checkRange() # To update self.rangeEnd (and should change self.inOffpeakRange)
-			else:
-				print("Warning : Fail to swnitch off all. We will try again on next loop.")
+			if not self._rangeChangeDone:
+				self.logger.debug("OffpeakStrategy : not offpeak, switchOffAll()")
+				if self.switchOffAll():
+					if cycleDuration>LOOP_DELAY_VIRTUAL and allowSleep:
+						self.offpeakHoursRanges.sleepUntillNextRange(now)
+						self.checkRange() # To update self.rangeEnd (and should change self.inOffpeakRange)
+					else:
+						self._rangeChangeDone = True
+						return self.offpeakHoursRanges.getTime2NextRange(now)
+				else:
+					self.logger.warning("Fail to switch off all. We will try again on next loop.")
+		return 0