+++ src/openhems/modules/network/network_helper.py
@@ -7,6 +7,7 @@ from openhems.modules.util import (
 )
 from .driver.home_assistant_api import HomeAssistantAPI
 from .driver.fake_network import FakeNetwork
+from .network import OpenHEMSNetwork
 
 def getNetworkFromConfiguration(logger, configurator:ConfigurationManager):
 	"""
@@ -21,5 +22,5 @@ def getNetworkFromConfiguration(logger, configurator:ConfigurationManager):
 		networkUpdater = FakeNetwork(configurator)
 	else:
 		raise ConfigurationException(f"Invalid server.network configuration '{networkSource}'")
-	network = networkUpdater.getNetwork(logger)
+	network = OpenHEMSNetwork(logger, networkUpdater, configurator.get("network.nodes"))
 	return network