+++ src/openhems/server.py
@@ -2,10 +2,11 @@
 """
 This is the server thread witch aim to centralize information and take right deccisions
 """
-import os
 import time
-from openhems.modules.energy_strategy import OffPeakStrategy
+import datetime
+from openhems.modules.energy_strategy import OffPeakStrategy, SwitchoffStrategy, LOOP_DELAY_VIRTUAL
 from openhems.modules.network import HomeStateUpdaterException
+from openhems.modules.util import CastUtililty
 from openhems.modules.util.configuration_manager import ConfigurationManager, ConfigurationException
 
 
@@ -19,26 +20,64 @@ class OpenHEMSServer:
 		self.logger = mylogger
 		self.network = network
 		self.loopDelay = serverConf.get("server.loopDelay")
-		strategy = serverConf.get("server.strategy").lower()
-		if strategy=="offpeak":
-			self.strategy = OffPeakStrategy(mylogger, self.network)
-		elif strategy=="emhass":
-			# pylint: disable=import-outside-toplevel
-			# Avoid to import EmhassStrategy and all it's dependances when no needs.
-			from openhems.modules.energy_strategy.emhass_strategy import EmhassStrategy
-			self.strategy = EmhassStrategy(mylogger, self.network, serverConf)
-		else:
-			self.logger.critical("OpenHEMSServer() : Unknown strategy '%s'", strategy)
-			os._exit(1)
+		strategies = serverConf.get("server.strategies")
+		self.strategies = []
+		throwErr = None
+		for strategyParams in strategies:
+			strategy = strategyParams.get("class", "")
+			strategyId = strategyParams.get("id", strategy)
+			if strategy=="offpeak":
+				self.strategies.append(OffPeakStrategy(mylogger, self.network, strategyId))
+			elif strategy=="switchoff":
+				offhoursrange = strategyParams.get('offrange', "[22h-6h]")
+				condition = strategyParams.get('condition', True)
+				reverse = CastUtililty.toTypeBool(strategyParams.get('reverse', False))
+				strategyObj = SwitchoffStrategy(mylogger, self.network, strategyId, offhoursrange, reverse, condition)
+				self.strategies.append(strategyObj)
+			elif strategy=="emhass":
+				# pylint: disable=import-outside-toplevel
+				# Avoid to import EmhassStrategy and all it's dependances when no needs.
+				from openhems.modules.energy_strategy.emhass_strategy import EmhassStrategy
+				self.strategies.append(EmhassStrategy(mylogger, self.network, serverConf, strategyId))
+			else:
+				msg = f"OpenHEMSServer() : Unknown strategy '{strategy}'"
+				self.logger.critical(msg)
+				throwErr = msg
+		if throwErr is not None:
+			raise ConfigurationException(throwErr)
+		self.allowSleep = len(self.strategies)==1
 
-	def loop(self, loopDelay):
+	def getSchedule(self):
+		"""
+		Return scheduled planning.
+		"""
+		schedule = {}
+		for strategy in self.strategies:
+			nodes = strategy.getSchedulableNodes()
+			for node in nodes:
+				myid = node.id
+				sc = node.getSchedule()
+				if sc is not None:
+					schedule[myid] = sc
+		return schedule
+
+	def loop(self, loopDelay, now=None):
 		"""
 		It's the content of each loop.
 		If loop delay=0, we consider that we never sleep (For test or reactivity).
 		"""
+		if now is None:
+			now = datetime.datetime.now()
 		self.logger.debug("OpenHEMSServer.loop()")
 		self.network.updateStates()
-		self.strategy.updateNetwork(loopDelay)
+		time2wait = 86400
+		allowSleep = self.allowSleep and loopDelay>LOOP_DELAY_VIRTUAL
+		for strategy in self.strategies:
+			t = strategy.updateNetwork(loopDelay, allowSleep, now)
+			time2wait = min(t, time2wait)
+		if allowSleep and time2wait > 0:
+			self.logger.info("Loop sleep(%d min)", round(time2wait/60))
+			time.sleep(time2wait)
 
 	def run(self, loopDelay=0):
 		"""